# docs/copilot_studio/openapi.json
{
  "components": {
    "schemas": {
      "CardAction": {
        "description": "What the card's Action.Submit posts back. Anything else in the body (an approver_id, say) is ignored.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "edited_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Content"
          },
          "proposal_id": {
            "title": "Proposal Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "action",
          "proposal_id"
        ],
        "title": "CardAction",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ProposalRef": {
        "properties": {
          "proposal_id": {
            "title": "Proposal Id",
            "type": "string"
          }
        },
        "required": [
          "proposal_id"
        ],
        "title": "ProposalRef",
        "type": "object"
      },
      "RiskQuery": {
        "properties": {
          "related_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Item Id"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "RiskQuery",
        "type": "object"
      },
      "RiskRef": {
        "properties": {
          "risk_id": {
            "title": "Risk Id",
            "type": "string"
          }
        },
        "required": [
          "risk_id"
        ],
        "title": "RiskRef",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "The shared secret (PM_COPILOT_API_KEY) the Power Platform connection sends. Set it once in the connector's security tab.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Proposal approvals as Adaptive Cards in Teams, and the risk log read conversationally. Import /openapi.json as a Power Platform custom connector's definition. Every action needs X-API-Key; the decision endpoints also need X-Authenticated-User, set by the flow from the Teams responder, never from the card.",
    "title": "P2 PM Delivery Steward -- Copilot Studio connector backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/card_action": {
      "post": {
        "description": "The card's Approve or Reject. WHO is acting is the X-Authenticated-User header, set by the flow from the Teams responder, never from the card or the body. Always answers 200 with an outcome: sent, rejected, refused, send_failed or held.",
        "operationId": "CardAction",
        "parameters": [
          {
            "in": "header",
            "name": "x-authenticated-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Authenticated-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cardaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Approve or reject from a card"
      }
    },
    "/explain_risk": {
      "post": {
        "description": "One risk explained: its tracker item, who has it, and what is still waiting for a decision about it. Reads only.",
        "operationId": "ExplainRisk",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskRef"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Explainrisk",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Explain a risk"
      }
    },
    "/get_decision_card": {
      "post": {
        "description": "The card to show once a decision is made: who decided, when, what was proposed, what was applied.",
        "operationId": "GetDecisionCard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRef"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getdecisioncard",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get the decision card"
      }
    },
    "/health": {
      "get": {
        "description": "No key needed. Says whether a key and approvers are configured (never what they are) and whether a post would reach Teams or only a log.",
        "operationId": "Health",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health check"
      }
    },
    "/list_pending_approvals": {
      "post": {
        "description": "Every proposal awaiting a decision, each with the Adaptive Card to show for it (a morning brief or end-of-day summary to approve, or a batch from a channel record to read and reject).",
        "operationId": "ListPendingApprovals",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Listpendingapprovals",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List pending approvals"
      }
    },
    "/list_risks": {
      "post": {
        "description": "The risk log, optionally narrowed by status, severity or tracker item, with a plain-language summary. Reads only.",
        "operationId": "ListRisks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Listrisks",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List risks"
      }
    }
  }
}